ルールが存在するアクセスルールグループの削除を制限

+++ plugins/bc-admin-third/templates/Admin/PermissionGroups/edit.php
@@ -69,16 +69,18 @@ $this->BcAdmin->addAdminMainBodyHeaderLinks([
     ]) ?>
   </div>
   <div class="bca-actions__sub">
-      <?= $this->BcAdminForm->postLink(
-        __d('baser_core', '削除'),
-        ['action' => 'delete', $userGroupId, $entity->id],
-        ['block' => true,
-          'confirm' => __d('baser_core', '{0} を本当に削除してもいいですか？', $entity->name),
-          'class' => 'bca-submit-token button bca-btn bca-actions__item',
-          'data-bca-btn-type' => 'delete',
-          'data-bca-btn-size' => 'sm'
-        ]
-      ) ?>
+      <?php if ($isDeletable): ?>
+        <?= $this->BcAdminForm->postLink(
+          __d('baser_core', '削除'),
+          ['action' => 'delete', $userGroupId, $entity->id],
+          ['block' => true,
+            'confirm' => __d('baser_core', '{0} を本当に削除してもいいですか？', $entity->name),
+            'class' => 'bca-submit-token button bca-btn bca-actions__item',
+            'data-bca-btn-type' => 'delete',
+            'data-bca-btn-size' => 'sm'
+          ]
+        ) ?>
+      <?php endif ?>
   </div>
 </div>
